+++ Example/Example controllers/Search/CoreDataSearchViewController.swift
@@ -30,8 +30,8 @@ class CoreDataSearchViewController: UIViewController, DTTableViewManageable {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        manager = DTTableViewManager(storage: CoreDataStorage(fetchedResultsController: fetchResultsController))
         manager.register(BankCell.self)
-        manager.storage = CoreDataStorage(fetchedResultsController: fetchResultsController)
         manager.tableViewUpdater?.didUpdateContent = { [weak self] _ in
             self?.tableView.isHidden = self?.tableView.numberOfSections == 0
         }